produce, combine, and publish coverage results
Issues:
Fixes #827

Problem: Much of the code coverage of agent code is produced by
tests run from this repository on a "controller" test node.  To observe
how our code coverage is evolving we need that data to be combined with
data from the agent repository.

Analysis:  This code depends on the code coverage data for the agent
already existing.  If it does (in the correct location:
/srv/nfs/openstack/testresults/coverage/agent/BRANCH/COMMIT/*) then this
change will add the driver's agent coverage to that location, combine
the full set of results, and publish them to coveralls.io

Tests: We know this change does not cause regressions because the smoke
tests pass.  Once this commit is merged, we can validate that coverage
results are updated with each test run.