Member ports owned by 'network:f5lbaasv2' are still being torn down in driver_v2.py
Issues:
Fixes #729

Problem:
Earlier this year, we made changes to all OpenStack releases of the
driver, to no longer manage ports for members. Most of our focus was on
port creation for those members, but some logic remains on member
deletion, to delete a port associated with a member only if we,
network:f5lbaasv2 own that port. This will most likely not cause a
problem of us inadvertently tearing down ports unless those ports have
that same owner as seen above. I think we can safely remove this code.

Analysis:
Removed the dead code in driver_v2.py

Tests:
Will run all tests, to determine if any side-effects are unaccounted
for.