Update port creation RPC methods to include vnic_type/binding_profile
Issues:
Fixes #849

Problem:
There needs to be a way for the agent to specify whether or not to
create 'baremetal' vs. 'normal' ports. Also, fix a couple of
issues with the way that we specify attributes for port creation.

Analysis:
The ML2 driver will only bind ports that are of type 'baremetal',
for the BIG-IP.  These changes add parameters to the port creation
RPC methods: vnic_type and binding_profile.

Fixes the way that the port attributes are assigned, including
using the passed in value for device_id.

Tests:
TBD