Add logic to rebind loadbalancers when the currently bound agent is dead/down
Issues:
Fixes #863

Problem:
In order for the OpenStack product to support agents that can come and go, we
need to be able to rebind the loadbalancer to an agent in the environment group.
We also need to have the scheduler selecting an agent consider the administrative
state of the agent and not select agents that report as admin_state_up False.

Analysis:

Add function to determine which agents are down in a given environment,
and remove them by rebinding the loadbalancers that are associated with
them.

1) Change neutron constants import to use neutron_lib
2) Add a callback to set the state of the agent.
3) Add a callback to get errored loadbalancers.
4) Changed format of many function calls so that the parameters are not listed
   one per line.
5) Update the loadbalancer getter callbacks to scrub dead agents and
   rebind associated loadbalancers.

Tests:
f5lbaasdriver/test/tempest/tests/api/