Implement required _handle_driver_error() for the LBaaSv2 driver
Issues:
Fixes #864

Problem:
Whenever an exception is thrown by neutron-lbaas for a driver operation
error the _handle_driver_error method is called on the provider driver.
We don't implement this and it causes another exception

Analysis:
Add function definition to the driver.  It does not perform any special
handling of the error.

Tests:
neutron_lbaas/test/tempest