device_id exception raised when creating Neutron network port (v10.3, HPB)
Issues:
Fixes #930

Problem: create_port_on_network() fails to set a device_id in
port data unless one is passed into the function. Because the
agent calls this method without a device_id it will always
fail.

Analysis: Code was recently changed so that the passed in device_id
is always used, but failed to account for situations when device_id
is not given. Added additional code to set device_id to a uuid ID
when no device_id is given.

Tests: