update __init__ version number
to be 112.1.0.
add 100 for the first number; so that we can easily know
the corresponding version. So this means 112.1.0 is going
to be used as the first potential driver release for
the so-called "official" H3C controller.

The version number can be discussed, especially when
more sdn controllers are being integrated. We can discuss
& modify & broadcast the version changes if that really
happens in the future.