avoid getting lb from db
Let's avoid modifying the schedule() function's prototype
as that might impact too many places.

To avoid unnecesary db call when the lb was already scheduled
before. This seems to be a safer way which modifies the least
code. Seems a little risky if we try to use the lb passed from
member.pool.loadbalancer because this content from the
third-party might differ.